backend/requirements.txt
# Core API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# OpenAI (without tiktoken to avoid build issues)
openai==1.3.5

# Data processing
pandas==2.1.3
python-dotenv==1.0.0

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Vector Database
chromadb==0.4.18
